sustainabilityhub/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
Pillow>=10.0.0
django-ckeditor>=6.7.0
django-ratelimit>=4.1.0
redis>=5.0.0
django-redis>=5.4.0
django-cors-headers>=4.3.0
django-filter>=23.3
django-debug-toolbar>=4.2.0
celery>=5.3.0
django-extensions>=3.2.0
